backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.13.0
pyserial==3.5
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6 